[metasrv] refactor: move raft config into metasrv/raft/config.rs
