+++ metasrv/src/raft/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod config;
 pub mod log;
 pub mod sled_key_spaces;
 pub mod state;
